+++ src/factorizations/diagonal.jl
@@ -1,6 +1,7 @@
 # DiagonalTensorMap
 # -----------------
 _repack_diagonal(d::DiagonalTensorMap) = Diagonal(d.data)
+_repack_diagonal(d::SectorVector) = Diagonal(parent(d))
 
 MAK.diagview(t::DiagonalTensorMap) = SectorVector(t.data, TensorKit.diagonalblockstructure(space(t)))
 
@@ -10,26 +11,6 @@ for f in (
         :eigh_trunc, :eigh_vals, :left_polar, :right_polar,
     )
     @eval MAK.copy_input(::typeof($f), d::DiagonalTensorMap) = copy(d)
-end
-
-for f! in (:eig_full!, :eig_trunc!)
-    @eval function MAK.initialize_output(
-            ::typeof($f!), d::AbstractTensorMap, ::DiagonalAlgorithm
-        )
-        return d, similar(d)
-    end
-end
-
-for f! in (:eigh_full!, :eigh_trunc!)
-    @eval function MAK.initialize_output(
-            ::typeof($f!), d::AbstractTensorMap, ::DiagonalAlgorithm
-        )
-        if scalartype(d) <: Real
-            return d, similar(d)
-        else
-            return similar(d, real(scalartype(d))), similar(d)
-        end
-    end
 end
 
 for f! in (:qr_full!, :qr_compact!)
@@ -72,16 +53,12 @@ function MAK.initialize_output(
     V_cod = fuse(codomain(t))
     V_dom = fuse(domain(t))
     U = similar(t, codomain(t) ← V_cod)
-    S = DiagonalTensorMap{real(scalartype(t))}(undef, V_cod ← V_dom)
+    S = similar_diagonal(t, real(scalartype(t)), V_cod)
     Vᴴ = similar(t, V_dom ← domain(t))
     return U, S, Vᴴ
 end
 
-for f! in
-    (
-        :qr_full!, :qr_compact!, :lq_full!, :lq_compact!, :eig_full!, :eig_trunc!, :eigh_full!,
-        :eigh_trunc!, :right_orth!, :left_orth!,
-    )
+for f! in (:qr_full!, :qr_compact!, :lq_full!, :lq_compact!, :right_orth!, :left_orth!)
     @eval function MAK.$f!(d::DiagonalTensorMap, F, alg::DiagonalAlgorithm)
         $f!(_repack_diagonal(d), _repack_diagonal.(F), alg)
         return F
@@ -93,17 +70,10 @@ function MAK.svd_compact!(t::AbstractTensorMap, USVᴴ, alg::DiagonalAlgorithm)
     return svd_full!(t, USVᴴ, alg)
 end
 
-# f_vals
-# ------
-for f! in (:eig_vals!, :eigh_vals!, :svd_vals!)
-    @eval function MAK.$f!(d::AbstractTensorMap, V, alg::DiagonalAlgorithm)
-        $f!(_repack_diagonal(d), diagview(_repack_diagonal(V)), alg)
-        return V
-    end
-    @eval function MAK.initialize_output(
-            ::typeof($f!), d::DiagonalTensorMap, alg::DiagonalAlgorithm
-        )
-        data = MAK.initialize_output($f!, _repack_diagonal(d), alg)
-        return DiagonalTensorMap(data, d.domain)
-    end
+# For diagonal inputs we don't have to promote the scalartype since we know they are symmetric
+function MAK.initialize_output(::typeof(eig_vals!), t::AbstractTensorMap, alg::DiagonalAlgorithm)
+    V_D = fuse(domain(t))
+    Tc = complex(scalartype(t))
+    A = similarstoragetype(t, Tc)
+    return SectorVector{Tc, sectortype(t), A}(undef, V_D)
 end